+++ vue/src/i18n/en-US/home.json
@@ -11,7 +11,8 @@
       "header": {
         "name": "Name",
         "status": "State",
-        "create": "Created"
+        "create": "Created",
+        "duration": "Runtime"
       }
     }
   },
